requirements.txt
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.17
django-debug-toolbar==2.2
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
libvirt-python==5.7.0
MarkupSafe==1.1.1
mysqlclient==2.0.3
numpy==1.19.5
packaging==20.8
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
uWSGI==2.0.19.1
websockify==0.9.0
